api.jsの作成
変更内容：APIリクエストをまとめたapi.jsを作成した。create_task_from_json.jsは削除した。またエラーハンドリングを実装した。
